Separated AnimationBase and AnimationUnknown classes into separate files.
