+++ Source/AliveLibAE/Animation.hpp
@@ -1,10 +1,10 @@
 #pragma once
 
+#include "AnimationBase.hpp"
 #include "FunctionFwd.hpp"
 #include "DynamicArray.hpp"
 #include "Psx.hpp"
 #include "BaseGameObject.hpp"
-#include "Primitives.hpp"
 #include "FixedPoint.hpp"
 
 using TFrameCallBackType = int(CC *)(void *, __int16 *);
@@ -80,110 +80,6 @@ struct AnimationHeader
 };
 //ALIVE_ASSERT_SIZEOF(AnimationHeader, 0x8);
 
-enum AnimFlags
-{
-    // Bit 1 = some unknown effect on frame ordering
-    eBit1 = 0x1,
-
-    // Bit 2 = Visible/render flag?
-    eBit2_Animate = 0x2,
-
-    // Bit 3 = nothing ?
-    eBit3_Render = 0x4,
-
-    // Bit 4 = Horizontal flip
-    eBit4 = 0x8,
-
-    // Bit 5 = Vertical flip
-    eBit5_FlipX = 0x10,
-
-    // Bit 6 = nothing ?
-    eBit6_FlipY = 0x20,
-
-    // Bit 7 = loop
-    eBit7_SwapXY = 0x40,
-
-    // Bit 8 = ?
-    eBit8_Loop = 0x80,
-
-    // Bit 9 = ?
-    eBit9 = 0x100,
-
-    // Bit 10 = ?
-    eBit10_alternating_flag = 0x200,
-
-    // Bit 11 = ?
-    eBit11_bToggle_Bit10 = 0x400,
-
-    // Bit 12 = prevents updating or gets anims stuck ??
-    eBit12_ForwardLoopCompleted = 0x800,
-
-    // Bit 13 = colour depth related
-    eBit13_Is8Bit = 0x1000,
-
-    // Bit 14 = transparency enabled
-    eBit14_Is16Bit = 0x2000,
-
-    // Bit 15 = disable RGB changes? Shadows no longer work, all muds look like abe ?
-    eBit15_bSemiTrans = 0x4000,
-
-    // Bit 16 = nothing ?
-    eBit16_bBlending = 0x8000,
-
-    // Bit 17 = is last frame? causes instance chiseling of muds
-    eBit17 = 0x10000,
-
-    // Bit 18 = backwards (or forwards?) frame order
-    eBit18_IsLastFrame = 0x20000,
-
-    // Bit 19 = offset Y by H?
-    eBit19_LoopBackwards = 0x40000,
-
-    // Bit 20 = nothing ?
-    eBit20_use_xy_offset = 0x80000,
-
-    // Bit 21 = Hang/crash
-    eBit21 = 0x100000,
-
-    // Bit 22 = nothing ?
-    eBit22_DeadMode = 0x200000,
-
-    // Bit 23 = nothing ?
-    eBit23 = 0x400000,
-
-    // Bit 24 = Display vram ?
-    eBit24 = 0x800000,
-
-    eBit25_bDecompressDone = 0x1000000,
-
-    // Bit 26-32 = nothing ?
-};
-
-class AnimationBase
-{
-public:
-    // TODO: Virtuals must be on the base type, yet there is only 1 vtable pointing to derived ?
-    virtual void vDecode_40AC90();
-
-    virtual void vRender_40B820(int xpos, int ypos, int** pOt, __int16 width, signed int height);
-    virtual void vCleanUp_40C630() = 0;
-   
-    // TODO: Restore vTable entry
-    //EXPORT virtual __int16 Animationv_40B200();
-    EXPORT virtual char Animation_v_40BEE0(__int16 a2, __int16 a3, int a4, __int16 a5, __int16 op1);
-
-    EXPORT static void CC AnimateAll_40AC20(DynamicArrayT<AnimationBase>* pAnimations);
-
-    BitField32<AnimFlags> field_4_flags;
-    BYTE field_8_r;
-    BYTE field_9_g;
-    BYTE field_A_b;
-    BYTE field_B_render_mode;
-    WORD field_C_render_layer;
-    WORD field_E_frame_change_counter;
-};
-ALIVE_ASSERT_SIZEOF(AnimationBase, 0x10);
-
 struct BanHeader
 {
     WORD mMaxW = 0;
@@ -206,7 +102,7 @@ struct FrameHeader
 class Animation : public AnimationBase
 {
 public:
-    EXPORT signed __int16 Set_Animation_Data_409C80(int frameTableOffset, BYTE **pAnimRes);
+    EXPORT signed __int16 Set_Animation_Data_409C80(int frameTableOffset, BYTE** pAnimRes);
     EXPORT void Animation_Pal_Free_40C4C0();
 
     bool EnsureDecompressionBuffer();
@@ -244,7 +140,7 @@ public:
     EXPORT FrameInfoHeader* Get_FrameHeader_40B730(__int16 frame);
     EXPORT void Get_Frame_Rect_409E10(PSX_RECT *pRect);
     EXPORT WORD Get_Frame_Count_40AC70();
-    EXPORT signed __int16 Init_40A030(int frameTableOffset, DynamicArray *animList, BaseGameObject *pGameObj, unsigned __int16 maxW, unsigned __int16 maxH, BYTE **ppAnimData, unsigned __int8 unknown1, signed int pal_depth, char unknown3);
+    EXPORT signed __int16 Init_40A030(int frameTableOffset, DynamicArray* animList, BaseGameObject* pGameObj, unsigned __int16 maxW, unsigned __int16 maxH, BYTE** ppAnimData, unsigned __int8 unknown1, signed int pal_depth, char unknown3);
     EXPORT void Get_Frame_Offset_40C480(__int16* pBoundingX, __int16* pBoundingY);
     EXPORT void Get_Frame_Width_Height_40C400(__int16* pWidth, __int16* pHeight);
 
@@ -258,24 +154,3 @@ inline bool IsLastFrame(const Animation* pAnim)
     const auto pHeader = reinterpret_cast<const AnimationHeader*>(&pAnimData[pAnim->field_18_frame_table_offset]);
     return (pAnim->field_92_current_frame == pHeader->field_2_num_frames - 1);
 }
-
-// TODO: Figure out how this differs from the standard Animation
-class AnimationUnknown : public AnimationBase
-{
-public:
-    virtual void vDecode_40AC90() override;
-    virtual void vRender_40B820(int xpos, int ypos, int** pOt, __int16 width, signed int height) override;
-    virtual void vCleanUp_40C630() override;
-
-    EXPORT void GetRenderedSize_40C980(PSX_RECT* pRect);
-
-private:
-    EXPORT void vRender_40C690(int xpos, int ypos, int** pOt, int width, int height);
-    EXPORT void vCleanUp_40C9C0();
-
-public:
-    Poly_FT4 field_10_polys[2];
-    Animation* field_68_anim_ptr;
-    FP field_6C_scale;
-};
-ALIVE_ASSERT_SIZEOF(AnimationUnknown, 0x70);
