+++ Source/AliveLibAE/BaseAnimatedWithPhysicsGameObject.hpp
@@ -4,6 +4,7 @@
 #include "FixedPoint.hpp"
 #include "FunctionFwd.hpp"
 #include "Animation.hpp"
+#include "AnimationUnknown.hpp"
 #include "Map.hpp"
 
 namespace Test
